Revert "Extract buffer logic to legacy adapter rollout fn (#171)"
This reverts commit 118b6a02b494cd8ffa53a3f21471ec5862561631.
